move disqualification calculator
