+++ app/services/calculators/addition_calculator.rb
@@ -10,26 +10,6 @@ module Calculators
       end
     end
 
-    #
-    # If disqualification and no length is given: conviction start date is spent date
-    #
-    class StartPlusZeroMonths < AdditionCalculator
-      def added_time
-        { months: 0 }
-      end
-
-      def conviction_length
-        { (disclosure_check.conviction_length_type || "no_length").to_sym => disclosure_check.conviction_length }
-      end
-
-      def expiry_date
-        return ResultsVariant::INDEFINITE if indefinite_length?
-
-        super
-        conviction_end_date.advance(added_time)
-      end
-    end
-
     # If length is given: conviction end date + 6 months
     # If no length is given: conviction start date + 24 months
     #
